Troubleshooting an issue with the migration of the managers.
